requirements.txt
matplotlib>=3.7.0
networkx>=3.0
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
reportlab>=4.0.0
requests>=2.28.0 